Fix CompactBigIntWriter when the value starts with a zero byte
The serialization of BigIntegers may include an extra zero byte as
prefix when the first byte is > 127, to act as a sign bit and distinguish
it from a negative value. We only deal with unsigned values, so we must
ignore that initial zero byte.

Ref: https://docs.oracle.com/javase/7/docs/api/java/math/BigInteger.html#toByteArray%28%29

This should fix https://github.com/emeraldpay/polkaj/issues/32